move strided drop layer to own file
